+++ server/src/controllers/postController/createPostController.ts
@@ -1,5 +1,5 @@
 import {Request,Response} from 'express'
-import reviewDatas from '../models/reviewDatas'
+import reviewDatas from '../../models/reviewDatas'
 
 const createPostController = async (req: Request,res: Response)=>{
     const newReviewDatas = new reviewDatas({
