+++ server/src/controllers/postController/getPostController.ts
@@ -1,5 +1,5 @@
 import {Request,Response} from 'express'
-import reviewDatas from '../models/reviewDatas'
+import reviewDatas from '../../models/reviewDatas'
 
 const getPostController = async (req: Request, res: Response) => {
         const datas = await reviewDatas.find();
